[finishes #171810827] fix continue on error in watcher
