+++ tests/integration/adam_test.go
@@ -1,7 +1,6 @@
 package integration
 
 import (
-	"errors"
 	"fmt"
 	"github.com/itmo-eve/eden/pkg/adam"
 	"github.com/itmo-eve/eden/pkg/cloud"
@@ -9,8 +8,7 @@ import (
 	"github.com/itmo-eve/eden/pkg/einfo"
 	"github.com/itmo-eve/eden/pkg/elog"
 	"github.com/itmo-eve/eden/pkg/utils"
-	"github.com/lf-edge/eve/api/go/config"
-	uuid "github.com/satori/go.uuid"
+	"github.com/lf-edge/eve/api/go/info"
 	"os"
 	"path"
 	"path/filepath"
@@ -84,48 +82,6 @@ func TestAdamOnBoard(t *testing.T) {
 	t.Fatal("Onboarding timeout")
 }
 
-func adamPrepare() (adamCtx *adam.AdamCtx, id *uuid.UUID, err error) {
-	currentPath, err := os.Getwd()
-	if err != nil {
-		return nil, nil, err
-	}
-	ip := os.Getenv("IP")
-	if len(ip) == 0 {
-		ip, err = utils.GetIPForDockerAccess()
-		if err != nil {
-			return nil, nil, err
-		}
-	}
-	port := os.Getenv("ADAM_PORT")
-	if len(port) == 0 {
-		port = "3333"
-	}
-	adamDir := os.Getenv("ADAM_DIST")
-	if len(adamDir) == 0 {
-		adamDir = path.Join(filepath.Dir(filepath.Dir(currentPath)), "dist", "adam")
-		if stat, err := os.Stat(adamDir); err != nil || !stat.IsDir() {
-			return nil, nil, err
-		}
-	}
-	ctx := adam.AdamCtx{
-		Dir: adamDir,
-		Url: fmt.Sprintf("https://%s:%s", ip, port),
-	}
-	cmdOut, err := ctx.DeviceList()
-	if err != nil {
-		return nil, nil, err
-	}
-	if len(cmdOut) > 0 {
-		devUUID, err := uuid.FromString(cmdOut[0])
-		if err != nil {
-			return nil, nil, err
-		}
-		return &ctx, &devUUID, nil
-	} else {
-		return nil, nil, errors.New("no device found")
-	}
-}
-
 func TestAdamSetConfig(t *testing.T) {
 	ctx, devUUID, err := adamPrepare()
 	if err != nil {
@@ -151,118 +107,21 @@ func TestAdamLogs(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	err = elog.LogWatchWithTimeout(ctx.GetLogsDir(devUUID), map[string]string{"devId": devUUID.String()}, elog.HandleFirst, 600)
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestBaseImage(t *testing.T) {
-	ctx, devUUID, err := adamPrepare()
-	if err != nil {
-		t.Fatal(err)
-	}
-	baseOSVersion := os.Getenv("EVE_BASE_VERSION")
-	if len(baseOSVersion) == 0 {
-		eveBaseRef := os.Getenv("EVE_BASE_REF")
-		if len(eveBaseRef) == 0 {
-			eveBaseRef = "4.10.0"
+	done := make(chan error)
+	go func() {
+		done <- elog.LogWatchWithTimeout(ctx.GetLogsDir(devUUID), map[string]string{"devId": devUUID.String()}, elog.HandleFirst, 600)
+	}()
+	go func() {
+		handler := func(item *elog.LogItem) bool {
+			done <- nil
+			return elog.HandleFirst(item)
 		}
-		zArch := os.Getenv("ZARCH")
-		if len(eveBaseRef) == 0 {
-			zArch = "amd64"
+		err := elog.LogLast(ctx.GetLogsDir(devUUID), map[string]string{"devId": devUUID.String()}, handler)
+		if err != nil {
+			done <- err
 		}
-		baseOSVersion = fmt.Sprintf("%s-%s", eveBaseRef, zArch)
-	}
-
-	dsId := "eab8761b-5f89-4e0b-b757-4b87a9fa93ec"
-
-	imageID := "1ab8761b-5f89-4e0b-b757-4b87a9fa93ec"
-
-	baseID := "22b8761b-5f89-4e0b-b757-4b87a9fa93ec"
-
-	imageName := path.Join(filepath.Dir(ctx.Dir), "images", "baseos.qcow2")
-
-	fi, err := os.Stat(imageName)
-	if err != nil {
-		t.Fatal(err)
-	}
-	size := fi.Size()
-
-	sha256sum, err := utils.SHA256SUM(imageName)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	cloudCxt := &cloud.CloudCtx{}
-	err = cloudCxt.AddDatastore(&config.DatastoreConfig{
-		Id:       dsId,
-		DType:    config.DsType_DsHttp,
-		Fqdn:     "http://mydomain.adam:8888",
-		ApiKey:   "",
-		Password: "",
-		Dpath:    "",
-		Region:   "",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	img := &config.Image{
-		Uuidandversion: &config.UUIDandVersion{
-			Uuid:    imageID,
-			Version: "4",
-		},
-		Name:      filepath.Base(imageName),
-		Sha256:    sha256sum,
-		Iformat:   config.Format_QCOW2,
-		DsId:      dsId,
-		SizeBytes: size,
-		Siginfo: &config.SignatureInfo{
-			Intercertsurl: "",
-			Signercerturl: "",
-			Signature:     nil,
-		},
-	}
-	err = cloudCxt.AddImage(img)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = cloudCxt.AddBaseOsConfig(&config.BaseOSConfig{
-		Uuidandversion: &config.UUIDandVersion{
-			Uuid:    baseID,
-			Version: "4",
-		},
-		Drives: []*config.Drive{{
-			Image:        img,
-			Readonly:     false,
-			Preserve:     false,
-			Drvtype:      config.DriveType_Unclassified,
-			Target:       config.Target_TgtUnknown,
-			Maxsizebytes: size,
-		}},
-		Activate:      true,
-		BaseOSVersion: baseOSVersion,
-		BaseOSDetails: nil,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	deviceCtx := device.CreateWithBaseConfig(*devUUID, cloudCxt)
-	deviceCtx.SetBaseOSConfig([]string{baseID})
-	b, err := deviceCtx.GenerateJsonBytes()
-	if err != nil {
-		t.Fatal(err)
-	}
-	configToSet := fmt.Sprintf("%s", string(b))
-	t.Log(configToSet)
-	res, err := ctx.ConfigSet(devUUID.String(), configToSet)
-	if err != nil {
-		t.Log(res)
-		t.Fatal(err)
-	}
-
-	err = elog.LogWatchWithTimeout(ctx.GetLogsDir(devUUID), map[string]string{"devId": devUUID.String(), "eveVersion": baseOSVersion}, elog.HandleFirst, 600)
-	if err != nil {
+	}()
+	if <-done != nil {
 		t.Fatal(err)
 	}
 }
@@ -272,7 +131,24 @@ func TestAdamInfo(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	err = einfo.InfoWatchWithTimeout(ctx.GetInfoDir(devUUID), map[string]string{"devId": devUUID.String()}, einfo.ZInfoDevSWFind, einfo.HandleFirst, 600)
+	done := make(chan error)
+
+	go func() {
+		err = einfo.InfoWatchWithTimeout(ctx.GetInfoDir(devUUID), map[string]string{"devId": devUUID.String()}, einfo.ZInfoDevSWFind, einfo.HandleFirst, 300)
+		done <- err
+	}()
+	go func() {
+		handler := func(im *info.ZInfoMsg, ds []*info.ZInfoDevSW) bool {
+			einfo.ZInfoDevSWPrn(im, ds)
+			done <- nil
+			return true
+		}
+		err = einfo.InfoLast(ctx.GetInfoDir(devUUID), map[string]string{"devId": devUUID.String()}, einfo.ZInfoDevSWFind, handler)
+		if err != nil {
+			done <- err
+		}
+	}()
+	err = <-done
 	if err != nil {
 		t.Fatal(err)
 	}
